[chain]fix uncle verify.
